// backend/src/dto/projects/create-project.dto.ts
export interface CreateProjectDto {
  organizationId: string;
  teamId?: string | null;
  name: string;
  description?: string | null;
  status?: 'PLANNING' | 'ACTIVE' | 'ON_HOLD' | 'COMPLETED' | 'ARCHIVED';
  priority?: 'LOW' | 'MEDIUM' | 'HIGH' | 'CRITICAL';
  startDate?: string | null;
  dueDate?: string | null;
}